refactor Transport to be an interface
To support replying OTLP events we need to use a different HTTP client than
what is in use right now.
This refactor changes Transport to be an interface with a single method that
is used to replay events.

This hides internal details on where the data is being replayed to.

This change is needed to account for the difference in implementation
between APM and OTLP: APM uses a single endpoint for all data types, OTLP
has separate endpoints for each data type.