add apm- prefix to agent names
EventHandler uses a single transport to replay events. This poses a limitation
on mixing APM and OTLP data during replay: they need different transports but
the current code flow does not allow that.

As a workaround to overcome the limitation, change the way we handle the empty
config.AgentName case: if empty, load all files starting with 'apm-'. By
applying this change the current behaviour is retained when adding support
for the additional OTLP transport.