Fix paths outside of the rootfs could be inadvertently produced
until v0.2.4 it was possible for certain rootfs and path combinations (in particular, where a malicious Unix-style `/-` separated unsafe path was used with a Windows-style rootfs path) to result in generated paths that were outside of the provided rootfs.

**Patches**
c121231e1276e11049547bee5ce68d5a2cfe2d9b is the patch fixing this issue. v0.2.4 contains the fix.