bench: Do not panic on `context.DeadlineExceeded` (#215)
This commit changes the behavior of the `apmbench` command when the
context deadline is exceeded. Instead of panicking, the error will be
ignored.

This change is necessary to avoid the panic when the context deadline is
exceeded, which is a common scenario when running the `apmbench` command
since the benchmark targets are usually pushed to their limits and some
requests are bound to eventually exceed the deadline.

Signed-off-by: Marc Lopez Rubio <marc5.12@outlook.com>